OAlg.AbelianGroup.Definition

Description

homomorphisms between finitely generated abelian groups.

Abelian Group

newtype AbGroup

finitely generate abelian group, i.e. the cartesian product of cyclic groups Z/n and are represented as a formal product with symbols in ZMod.

Definition Let g be in AbGroup. We call g smith normal if and only if there exists a sequence n 0, n 1 .. n (k-1) in N with length k and a exponent r in N such that:

2 <= n i for all 0 <= i < k.

n (i + 1) mod n i == 0 for all i in 0 <= i < k-1.

g == abg (n 0) * abg (n 1) * .. * abg (n (k-1)) * abg 0 ^ r.

Theorem Every finitely generated abelian group is isomorphic to a group in smith normal form. This isomorphism is given by isoSmithNormal.

Examples Finitely generated abelian groups constructed via abg and its multiplicative structure:

>>> abg 12
AbGroup[Z/12]

represents the cyclic group Z/12.

>>> abg 2 * abg 3
AbGroup[Z/2*Z/3]

represents the cartesian product of the groups Z/2 and Z/3.

>>> abg 6 * abg 4 * abg 4
AbGroup[Z/6*Z/4^2]

represents the cartesian product of the groups Z/6, Z/4 and Z/4.

>>> abg 0 ^ 6
AbGroup[Z^6]

represents the free abelian group Z ^ 6 of dimension 6.

>>> one () :: AbGroup
AbGroup[]

represents the cartesian product of zero cyclic groups and

>>> one () * abg 4 * abg 6 == abg 4 * abg 6
True

Examples Checking for smith normal via isSmithNormal:

>>> isSmithNormal (abg 4)
True

>>> isSmithNormal (abg 2 * abg 2)
True

>>> isSmithNormal (abg 17 * abg 51)
True

>>> isSmithNormal (abg 2 * abg 4 * abg 0 ^ 3)
True

>>> isSmithNormal (abg 5 * abg 3)
False

>>> isSmithNormal (abg 0 * abg 3 * abg 6)
False

>>> isSmithNormal (abg 1 * abg 4)
False

>>> isSmithNormal (one ())
True

Examples The associated isomorphism in AbHom of a finitely generated abelian group given by isoSmithNormal.

>>> end (isoSmithNormal (abg 3 * abg 5))
AbGroup[Z/15]

>>> end (isoSmithNormal (abg 2 * abg 4 * abg 2))
AbGroup[Z/2^2*Z/4]

>>> end (isoSmithNormal (abg 4 * abg 6))
AbGroup[Z/2*Z/12]

>>> end (isoSmithNormal (abg 1))
AbGroup[]

Constructors

AbGroup (ProductSymbol ZMod)

abg :: N -> AbGroup

the cyclic group of the given order as a finitely generated abelian group.

isSmithNormal :: AbGroup -> Bool

checks if the given group is smith normal (see definition AbGroup).

abgDim :: AbGroup -> Dim' ZModHom

the associated dimension for matrices of ZModHom.

Homomorphism

newtype AbHom

additive homomorphism between finitely generated abelian groups which are represented by matrices over ZModHom.

Constructors

AbHom (Matrix ZModHom)

abh :: Orientation AbGroup -> [(ZModHom, N, N)] -> AbHom

the additive homomorphism with the given orientation and ZModHom-entries.

abh' :: Orientation AbGroup -> [(Z, N, N)] -> AbHom

the additive homomorphism with the given orientation and Z-entries.

abhz :: AbHom -> Matrix Z

the underlying Z-matrix.

zabh :: Matrix Z -> AbHom

the associated homomorphism between products of abg 0 given by the column - respectively row - length.

Adjunction

abhFreeAdjunction :: Adjunction AbHomFree (Matrix Z) AbHom

the projection AbHomFree as left adjoint.

data AbHomFree x y where

projection homomorphisms to Matrix Z.

Constructors

AbHomFree :: AbHomFree AbHom (Matrix Z)

FreeAbHom :: AbHomFree (Matrix Z) AbHom

Limes

abhProducts :: Products n AbHom

products for AbHom.

abhSums :: Sums n AbHom

sums for AbHom.

Generator

abgGeneratorTo :: AbGroup -> Generator To AbHom

the generator for a finitely generated abelian group.

Property Let a be in AbGroup, then holds a == g where Generator (DiagramChainTo g _) _ _ _ _ = abgGeneratorTo a.

X

xAbHom :: Q -> Orientation AbGroup -> X AbHom

random variable for AbHom given by a density and an orientation.

xAbHomTo :: Q -> N -> N -> N -> AbGroup -> X AbHom

random variable of homomorphisms between abelian groups with end equal to the given one.

   r s t
  [f    ] a
  [     ] b
  [g h  ] c

xAbHomFrom :: Q -> N -> N -> N -> AbGroup -> X AbHom

random variable of homomorphisms between abelian groups with start equal to the given one.

   a b c
  [f    ] r
  [g   l] s
  [h    ] t

stdMaxDim :: N

the maximal length of abelian groups for the standard random variable of type X AbGroup.

Property 1 <= stdMaxDim.

Proposition

prpAbHom :: Statement

validity of the algebraic structure of AbHom.