OAlg.AbelianGroup.ZMod

Description

Homomorphisms between cyclic groups.

A group is called cyclic if it is generated by one element. Such a group G is obviously abelian and isomorphic to Z/nZ for some n :: N, which is called its order. The group homomorphisms between cyclic groups form a Z-algebraic structure which is presented here.

Cyclic Group

newtype ZMod

cyclic group Z/n, i.e. the quotient group of Z divided by nZ.

Constructors

ZMod N

zmOrd :: ZMod -> N

order of the cyclic group.

Homomorphism

data ZModHom

additive homomorphisms between cyclic groups ZMod constructable via zmh.

Note The homomorphisms between two cyclic groups a and b form again a cyclic group which will be denoted by ZModHom (a:>b) (see zmhGenOrd).

toZ :: ZModHom -> Z

the underlying integer.

fromZ :: Z -> ZModHom

the endomorphism in ZMod 0 given by the integer, i.e. the inverse of toZ.

zmh :: Orientation ZMod -> Z -> ZModHom

the induced homomorphism.

Property Let a, b be in ZMod and r in Z then holds:

zmh a b r is valid if and only if zmhEligible a b r is true.

If zmhEligible a b r is false then the evaluation of zmh a b r will end up by throwing a NotEligible-exception.

zmhEligible :: ZMod -> ZMod -> Z -> Bool

predicate to determine the eligibility of a value in Z to induce a homomorphism between the given cyclic groups.

Property Let a = ZMod a', b = ZMod b' be in ZMod and r in Z, then holds: zmhEligible a b r is true if and only if mod0 (r * inj a') b' == 0.

zmhGenOrd :: Orientation ZMod -> (ZModHom, N)

zmhGenOrd (a:>b) = (r,o) where r is a generator for ZModHom(a:>b) with order o.

Note It follows that inj o ! r == zero (a:>b) and ZModHom(a:>b) is isomorphic to Z/o which is represented by ZMod o.

zmhGenerator :: Orientation ZMod -> ZModHom

zmhGenerator (a:>b) = h is a generator for the abelian group ZModHom(a:>b).

X

Since the algebraic structure of ZModHom is not balanced - which means that between any two cyclic groups there may be no nontrivial homomorphisms - it is appropriate to work with XOrtSite instead of XOrtOrientation for random variables with values in ZModHom.

xZModHom :: X Z -> Orientation ZMod -> X ZModHom

random variable for homomorphisms for the given orientation of ZMod.

xZModHomTo :: N -> X Z -> X ZMod -> XOrtSite To ZModHom

random variable for homomorphisms based on xZModTo.

xZModHomFrom :: N -> X Z -> X ZMod -> XOrtSite From ZModHom

random variable for homomorphisms based on xZModFrom.

xZModTo :: N -> ZMod -> X ZMod

random variable of cyclic groups admitting nontrivial homomorphisms to the given group.

xZModFrom :: N -> ZMod -> X ZMod

random variable of cyclic groups admitting non trivial homomorphisms from the given one.

Proposition

prpZModHom :: Statement

validity of the Z-algebraic structure of ZModHom.

Exception

data ZModException

exceptions for cyclic groups which are sub exceptions from SomeOAlgException.

Constructors

NotEligible String